+++ resolvers/index.js
@@ -1,0 +1,9 @@
+const Query = require('./Query');
+const Product = require('./Product');
+const User = require('./User');
+
+module.exports = {
+    Query,
+    Product,
+    User,
+}
